# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "operationId": "cf7da4e9f465af3d2496c970399c6882",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "username",
                                    "country_code",
                                    "mobile_number",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "uniq username without space and unexpected chars",
                                        "type": "string"
                                    },
                                    "country_code": {
                                        "description": "user country code without ziro. exp: 98",
                                        "type": "integer"
                                    },
                                    "mobile_number": {
                                        "description": "user mobile number with standard format without country_code and ziro. exp: 9102403100",
                                        "type": "integer"
                                    },
                                    "password": {
                                        "description": "8-30 char string password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful registration",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "send sms successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "خطای اعتبارسنجی فیلدها"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/user/verify-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify OTP",
                "description": "/**",
                "operationId": "205dacae9ae40de43d0ec0468cd9b936",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "country_code",
                                    "mobile_number",
                                    "code"
                                ],
                                "properties": {
                                    "country_code": {
                                        "description": "user country code without ziro. exp: 98",
                                        "type": "string"
                                    },
                                    "mobile_number": {
                                        "description": "user mobile number with standard format without country_code and ziro. exp: 9102403100",
                                        "type": "string"
                                    },
                                    "code": {
                                        "description": "otp code received from sms",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "$ref": "#/components/schemas/UserResource"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "Bearer token"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "example": "2024-06-24 12:34:56"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid OTP code",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid Code"
                                        },
                                        "errors": {
                                            "properties": {
                                                "code": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Invalid Code"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user with mobile number",
                "operationId": "046dca4b172e6aecf4b21a21fbb9d909",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "country_code",
                                    "mobile_number"
                                ],
                                "properties": {
                                    "country_code": {
                                        "description": "Country code of the user",
                                        "type": "string"
                                    },
                                    "mobile_number": {
                                        "description": "Mobile number of the user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        },
                                        "errors": {
                                            "properties": {
                                                "credentials": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Invalid mobile number or password"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/user/resend-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend OTP code",
                "operationId": "b34e4d783e7b886a7d6c83eb21fa6b3f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "country_code",
                                    "mobile_number"
                                ],
                                "properties": {
                                    "country_code": {
                                        "description": "Country code of the user",
                                        "type": "string"
                                    },
                                    "mobile_number": {
                                        "description": "Mobile number of the user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP resent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "OTP has been resent successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        },
                                        "errors": {
                                            "properties": {
                                                "credentials": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Invalid mobile number or country code"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/api/user/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout the user",
                "description": "Log the user out.",
                "operationId": "b96a6adaebf3c55981b38fbb8521e61a",
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized - User not logged in"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "username": {
                        "type": "string",
                        "example": "john_doe"
                    },
                    "country_code": {
                        "type": "string",
                        "example": "98"
                    },
                    "mobile_number": {
                        "type": "string",
                        "example": "9123456789"
                    },
                    "password": {
                        "type": "string",
                        "example": "12345678"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-06-24 12:34:56"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-06-24 12:34:56"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use a Bearer token to access these endpoints",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}